requirements.txt
# Core Dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Reddit API
praw>=7.7.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# AI/LLM
openai>=1.0.0

# Audio Processing
pydub>=0.25.1
wave

# Video Processing  
moviepy>=1.0.3
pillow>=10.0.0

# Async Support
aiohttp>=3.8.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.2

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0